Use same time source as kernel synproxy module.
Some 6 years ago the kernel TCP timestamping moved from using jiffies to a 1ms
clock. Up to now this module still used jiffies which are different precision
unless kernel is build with HZ=1000.

I'm not sure whether this actually breaks initial RTT measurement or not but
it should be safer to use the same clock precision as the rest of the kernel.