Handle the MSS exactly the same as current synproxy version.
As the RTC says the MSS can be used completely independently in each direction
of data flow. The result may be quite different maximum sizes in the two
directions.

The MSS sent by the client should be encoded in the SYNCOOKIE for later
reconstruction of syn queue entry. Our own configured MSS should be sent
in TCP options to client.

Note that when --mss option is not specified in filter configuration the MSS
option flag parsed from incoming packet is reset and SYN/ACK response does
not contain the option at all.