Fix double free of a skb buffer in the send direct mode.
In the --senddirect mode an explicit free of the buffer was called in case
of xmit failure. But the called function dev_queue_xmit explicitely states that

  "Regardless of the return value, the skb is consumed".

Under heavy load when the system reaches a packet rate high enough to exhaust
the interface queue capacity an error is returned from the xmit function.
Causing the caller to performe a double free.